refactor(api): splitted harvest.js in multiple files
