+++ api/lib/services/processors/harvest/HarvestError.js
@@ -1,0 +1,7 @@
+module.exports = class HarvestError extends Error {
+  constructor(message, options) {
+    super(message);
+    this.type = options?.type;
+    this.cause = options?.cause;
+  }
+};
